==> notebooks/0.2-mcr-read_sql_notebook.ipynb <==
## Reading the SQL database

import pandas as pd
import numpy as np

## The Target variable

target = pd.read_csv('../data/sql_credit_risk_target.csv', index_col=0).set_index('SK_ID_CURR')

target.head()

train = target.dropna()

train.head()

test = target.reindex(target.index.difference(train.index))

test.head()

train.to_csv('../data/processed/train_target.csv')
test.to_csv('../data/processed/test_target.csv')

## Credit Risk Features

df = pd.read_csv('../data/sql_credit_risk_features.csv', index_col=0).set_index('SK_ID_CURR')

df.columns

df.shape

df.columns

df.describe()

df.nunique()

df.isnull().sum()

target.head()

a = pd.merge(target, df, right_index=True, left_index=True, how='left')

a.shape

a.to_csv('../data/processed/full_sql.csv')
