==> notebooks/0.4-mcr-dropped_cols.ipynb <==
## Dropped Columns

import pandas as pd
import re

%matplotlib inline

low_count = ['AMT_REQ_CREDIT_BUREAU_DAY_y', 'APARTMENTS_MODE_y', 
             'FLOORSMAX_MEDI_y', 'NAME_TYPE_SUITE_y', 'NONLIVINGAREA_MEDI_y']
    # These have not so many nans but a lot for the target variable
not_so_many= ['EXT_SOURCE_3','YEARS_BEGINEXPLUATATION_AVG_y','FLOORSMAX_AVG_y']


csvs = pd.read_csv('../data/master.csv', index_col='SK_ID_CURR').drop(columns='Unnamed: 0') 
sql = pd.read_csv('../data/processed/full_sql.csv', index_col='SK_ID_CURR')
api = pd.read_csv('../data/api_data_clean.csv', index_col='sk_id_curr')

temp_1 = pd.merge(sql, csvs, how='left', right_index=True, left_index=True)

master = pd.merge(temp_1, api, how='left', right_index=True, left_index=True)

master[['EXT_SOURCE_3']].describe()

a = []
for col in master.columns:
    print(col)
    a.append(col)

df = pd.read_csv('../data/processed/master_2.csv', index_col=0)

df.columns.sort_values()

## EXT SOURCE 

master['EXT_SOURCE_2'] = pd.to_numeric(master['EXT_SOURCE_2'].apply(lambda x: x.split('%')[0]), errors='coerce') / 100

master['EXT_SOURCE'] = master[['EXT_SOURCE_1', 'EXT_SOURCE_2', 'EXT_SOURCE_3']].mean(axis=1)

master['EXT_SOURCE'] = master['EXT_SOURCE'].fillna(master['EXT_SOURCE'].mean())

df['EXT_SOURCE'] = master['EXT_SOURCE']

df = df.drop(columns='EXT_SOURCE_2')

## Flags

flags = []
pattern = re.compile("FLAG|flag")
for col in df.columns:
    if pattern.search(col):
        flags.append(col)

df['total_flags'] = df[flags].sum(axis=1) ** 2

df.to_csv('../data/processed/master_3.csv')

df.head()

## Part 4


df = df.drop(columns=['DEF_30_CNT_SOCIAL_CIRCLE', 'obs_30_cnt_social_circle'])

df['SOCIAL'] = df['DEF_60_CNT_SOCIAL_CIRCLE'].multiply(df['OBS_60_CNT_SOCIAL_CIRCLE'])

## Moar

df['LIVINGAREA_AVG'].apply(lambda x: None if x == -99999 else x).isnull().sum()

df = df.drop(columns=['LIVINGAREA_AVG'])

df['income_perc'] = df['AMT_ANNUITY'] / df['AMT_INCOME_TOTAL']

df['income_perc']

df.head()

df.to_csv('../data/processed/master_4.csv')
