==> notebooks/01-RF-QueryApi.ipynb <==
import requests
import pandas as pd

df = pd.DataFrame()
df_list = []
for i in range(1,51):
    url = 'https://api.hackathon-03.batch2.lisbondatascience.org/api/data?page_size=2000&page='+ str(i)
    payload = {'Authorization': 'token fe0bfd98e3920dd027bd4ca2f796b322e6f95c35'}
    result = requests.get(url,headers=payload).json()
    df_list.append(pd.DataFrame(result['results']))


merged_df = pd.concat(df_list)

merged_df.to_csv('../data/api_data.csv')
