==> notebooks/02-RF-CleaningApiData.ipynb <==
%load_ext autoreload

import pandas as pd
%autoreload 2
from utils import clean_api_df
% matplotlib inline 

df = pd.read_csv('../data/api_data.csv')
columns_to_drop = ['flag_document_12','Unnamed: 0',
                   'amt_req_credit_bureau_qrt','apartments_avg',
                   'emergencystate_mode']
df = clean_api_df(df,drop_columns=columns_to_drop)

df.nunique()

df.isnull().sum()

df.to_csv('../data/api_data_clean.csv',index=False)
