==> notebooks/03-RF-TrainClassifier.ipynb <==
%load_ext autoreload

import pandas as pd
from sklearn.ensemble import RandomForestClassifier
%autoreload 2
from utils import join_target_api,join_to_sql,prepare_dataset,clean_csv_data
% matplotlib inline 
from sklearn.model_selection import train_test_split
from sklearn.metrics import roc_auc_score
from xgboost import XGBClassifier

df = pd.read_csv('../data/processed/master_4.csv',index_col='SK_ID_CURR')

df.head()
train = df[~df.TARGET.isnull()]
test = df[df.TARGET.isnull()]

x_train,x_test,y_train,y_test = train_test_split(train.drop(columns=['TARGET']),
                                                 train['TARGET'],
                                                 random_state = 22
                                                )

clf = XGBClassifier(n_estimators=500,
                    n_jobs = -1,
                    random_state = 22
                    )

clf.fit(x_train,y_train)

results = clf.predict_proba(x_test)[:,1]

roc_auc_score(y_test,results)

r = pd.Series(data=results,index=test.index)

r.to_csv('submission3.csv')
